Add KMS support for S3 and Kinesis
Update the templates to allow you to specify an optional KMS key to use with the Exported reports S3 bucket and an optional KMS key to use with the Agent event Kinesis stream. If a KMS key ARN is specified, the templates will grant the KMS permissions required to the AWS objects that need to read from that encrypted source.

The KMS key used to encrypt the Exported reports S3 bucket can be different than the KMS key used to encrypt the Agent event Kinesis stream.

Note: a KMS key ARN is required only if the S3 bucket or Kinesis stream is configured to use server-side encryption with a customer-managed CMK. If server-side encryption is not used or an AWS-managed CMK is used, no extra permissions are required.